Tweak 64bit type.

+++ include/KHR/khrplatform.h
@@ -179,11 +179,11 @@ typedef unsigned __int64        khronos_uint64_t;
 typedef int                     khronos_int32_t;
 typedef unsigned int            khronos_uint32_t;
 #if defined(__arch64__) || defined(_LP64)
-typedef long int                khronos_int64_t;
-typedef unsigned long int       khronos_uint64_t;
+typedef int64_t                khronos_int64_t;
+typedef uint64_t       khronos_uint64_t;
 #else
-typedef long long int           khronos_int64_t;
-typedef unsigned long long int  khronos_uint64_t;
+typedef int64_t           khronos_int64_t;
+typedef uint64_t  khronos_uint64_t;
 #endif /* __arch64__ */
 #define KHRONOS_SUPPORT_INT64   1
 #define KHRONOS_SUPPORT_FLOAT   1
